♻️ (Profile Button) - Refactor profile button into modular components + Split the monolithic ProfileButton component into smaller, focused components for better maintainability
