--- src/components/Auth/ProfileButton.tsx
@@ -1,550 +1,0 @@
-import useAuthStore from "@/store/authStore";
-import {
-  AlertCircleIcon,
-  CheckCircleIcon,
-  EyeIcon,
-  EyeOffIcon,
-  InfoIcon,
-  SaveIcon,
-  TrashIcon,
-  UserIcon,
-  MailIcon,
-  KeyIcon,
-} from "lucide-react";
-import React, { useState } from "react";
-import { Alert, AlertDescription } from "../ui/alert";
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-} from "../ui/alert-dialog";
-import { Button } from "../ui/button";
-import { Dialog, DialogContent, DialogHeader, DialogTitle } from "../ui/dialog";
-import { Input } from "../ui/input";
-import { Label } from "../ui/label";
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "../ui/tabs";
-
-const ProfileButton: React.FC = () => {
-  const { user, updatePassword, updateEmail, deleteAccount } = useAuthStore();
-  const [isOpen, setIsOpen] = useState(false);
-  const [email, setEmail] = useState("");
-  const [currentPassword, setCurrentPassword] = useState("");
-  const [newPassword, setNewPassword] = useState("");
-  const [confirmPassword, setConfirmPassword] = useState("");
-  const [isLoading, setIsLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-  const [success, setSuccess] = useState<string | null>(null);
-  const [showCurrentPassword, setShowCurrentPassword] = useState(false);
-  const [showNewPassword, setShowNewPassword] = useState(false);
-  const [showConfirmPassword, setShowConfirmPassword] = useState(false);
-  const [activeTab, setActiveTab] = useState<"email" | "password" | "delete">("email");
-  const [isDeleteDialogOpen, setIsDeleteDialogOpen] = useState(false);
-  const [deletePassword, setDeletePassword] = useState("");
-  const [showDeletePassword, setShowDeletePassword] = useState(false);
-  const [deleteError, setDeleteError] = useState<string | null>(null);
-  const [deleteConfirmation, setDeleteConfirmation] = useState("");
-
-  // Réinitialiser les champs lorsque la modal est fermée ou l'utilisateur change
-  React.useEffect(() => {
-    if (!isOpen) {
-      setEmail(user?.email || "");
-      setCurrentPassword("");
-      setNewPassword("");
-      setConfirmPassword("");
-      setDeletePassword("");
-      setDeleteConfirmation("");
-      setError(null);
-      setSuccess(null);
-      setDeleteError(null);
-      setShowCurrentPassword(false);
-      setShowNewPassword(false);
-      setShowConfirmPassword(false);
-      setShowDeletePassword(false);
-      setIsDeleteDialogOpen(false);
-      setIsLoading(false); // Réinitialiser l'état de chargement
-      setActiveTab("email"); // Réinitialiser l'onglet actif à l'email
-    }
-  }, [isOpen, user]);
-
-  // Initialiser l'email avec celui de l'utilisateur lorsque la modal s'ouvre
-  React.useEffect(() => {
-    if (isOpen && user) {
-      setEmail(user.email || "");
-    }
-  }, [isOpen, user]);
-  
-  // Réinitialiser l'état de chargement lorsque l'utilisateur change
-  React.useEffect(() => {
-    // Si l'utilisateur change (déconnexion/reconnexion), réinitialiser l'état de chargement
-    setIsLoading(false);
-    setDeleteError(null);
-    setDeleteConfirmation("");
-  }, [user]);
-
-  // Si l'utilisateur n'est pas connecté, ne rien afficher
-  if (!user) return null;
-
-  const handleUpdateEmail = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setIsLoading(true);
-    setError(null);
-    setSuccess(null);
-
-    try {
-      // Vérifier que le mot de passe actuel est fourni
-      if (!currentPassword) {
-        setError(
-          "Veuillez entrer votre mot de passe actuel pour confirmer le changement d'email."
-        );
-        setIsLoading(false);
-        return;
-      }
-
-      // Vérifier que l'email est différent de l'email actuel
-      if (email === user?.email) {
-        setError(
-          "La nouvelle adresse email est identique à l'adresse actuelle."
-        );
-        setIsLoading(false);
-        return;
-      }
-
-      // Mettre à jour l'email avec la fonction du store
-      const { error } = await updateEmail(email, currentPassword);
-
-      if (error) {
-        setError(error);
-      } else {
-        setSuccess(
-          "Un email de confirmation a été envoyé à votre nouvelle adresse email. " +
-            "Veuillez cliquer sur le lien dans cet email pour confirmer le changement. " +
-            "IMPORTANT : Vous devez continuer à utiliser votre ancienne adresse email pour vous connecter jusqu'à ce que vous ayez confirmé le changement. " +
-            "Après avoir cliqué sur le lien de confirmation, veuillez vous déconnecter et vous reconnecter avec votre nouvelle adresse email."
-        );
-        // Réinitialiser le mot de passe actuel après succès
-        setCurrentPassword("");
-      }
-    } catch (err) {
-      setError("Une erreur est survenue.");
-      console.error(err);
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  const handleUpdatePassword = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setIsLoading(true);
-    setError(null);
-    setSuccess(null);
-
-    try {
-      // Vérifier que les mots de passe correspondent
-      if (newPassword !== confirmPassword) {
-        setError("Les mots de passe ne correspondent pas.");
-        setIsLoading(false);
-        return;
-      }
-
-      // TODO: Vérifier le mot de passe actuel avant de permettre la mise à jour
-      // Pour l'instant, on met à jour directement
-      const { error } = await updatePassword(newPassword);
-
-      if (error) {
-        setError(error);
-      } else {
-        setSuccess("Mot de passe mis à jour avec succès.");
-        setNewPassword("");
-        setConfirmPassword("");
-        setCurrentPassword("");
-      }
-    } catch (err) {
-      setError("Une erreur est survenue.");
-      console.error(err);
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  const handleDeleteAccount = async () => {
-    setIsLoading(true);
-    setDeleteError(null);
-
-    try {
-      if (!deletePassword) {
-        setDeleteError("Veuillez entrer votre mot de passe pour confirmer la suppression.");
-        setIsLoading(false);
-        return;
-      }
-
-      // Vérifier que l'utilisateur a saisi "Supprimer mes données" dans le champ de confirmation (insensible à la casse)
-      if (deleteConfirmation.toLowerCase() !== "supprimer mes données".toLowerCase()) {
-        setDeleteError("Veuillez saisir 'Supprimer mes données' pour confirmer.");
-        setIsLoading(false);
-        return;
-      }
-
-      const { error } = await deleteAccount(deletePassword);
-
-      if (error) {
-        setDeleteError(error);
-        setIsLoading(false);
-      } else {
-        // La suppression a réussi, la déconnexion est gérée dans la fonction deleteAccount
-        setIsDeleteDialogOpen(false);
-        setIsOpen(false);
-      }
-    } catch (err) {
-      setDeleteError("Une erreur est survenue lors de la suppression du compte.");
-      console.error(err);
-      setIsLoading(false);
-    }
-  };
-
-  return (
-    <>
-      <Button
-        variant="outline"
-        size="sm"
-        onClick={() => setIsOpen(true)}
-        className="flex items-center gap-2"
-      >
-        <UserIcon className="size-4" />
-        Profil
-      </Button>
-
-      <Dialog open={isOpen} onOpenChange={setIsOpen}>
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>Profil utilisateur</DialogTitle>
-          </DialogHeader>
-
-          <Tabs
-            defaultValue="email"
-            value={activeTab}
-            onValueChange={(value) => {
-              setActiveTab(value as "email" | "password" | "delete");
-              setError(null);
-              setSuccess(null);
-            }}
-            className="w-full"
-          >
-            <TabsList className="w-full grid grid-cols-3 mb-4">
-              <TabsTrigger value="email" className="flex items-center gap-1">
-                <MailIcon className="size-4" />
-                <span className="hidden sm:inline">Modifier l'email</span>
-                <span className="sm:hidden">Email</span>
-              </TabsTrigger>
-              <TabsTrigger value="password" className="flex items-center gap-1">
-                <KeyIcon className="size-4" />
-                <span className="hidden sm:inline">Modifier le mot de passe</span>
-                <span className="sm:hidden">Mot de passe</span>
-              </TabsTrigger>
-              <TabsTrigger value="delete" className="flex items-center gap-1 text-destructive">
-                <TrashIcon className="size-4" />
-                <span className="hidden sm:inline">Supprimer mes données</span>
-                <span className="sm:hidden">Supprimer</span>
-              </TabsTrigger>
-            </TabsList>
-
-            <TabsContent value="email" className="mt-0">
-              <form onSubmit={handleUpdateEmail} className="space-y-4">
-              <div className="space-y-2">
-                <Label htmlFor="email">Nouvel email</Label>
-                <Input
-                  id="email"
-                  type="email"
-                  value={email}
-                  onChange={(e) => setEmail(e.target.value)}
-                  placeholder="votre@email.com"
-                  required
-                />
-              </div>
-
-              <div className="space-y-2">
-                <Label htmlFor="currentPasswordEmail">
-                  Mot de passe actuel
-                </Label>
-                <div className="relative">
-                  <Input
-                    id="currentPasswordEmail"
-                    type={showCurrentPassword ? "text" : "password"}
-                    value={currentPassword}
-                    onChange={(e) => setCurrentPassword(e.target.value)}
-                    placeholder="••••••••"
-                    required
-                  />
-                  <button
-                    type="button"
-                    className="absolute right-3 top-1/2 -translate-y-1/2 text-gray-500 hover:text-gray-700 focus:outline-none"
-                    onClick={() => setShowCurrentPassword(!showCurrentPassword)}
-                    tabIndex={-1}
-                  >
-                    {showCurrentPassword ? (
-                      <EyeIcon className="size-4" />
-                    ) : (
-                      <EyeOffIcon className="size-4" />
-                    )}
-                  </button>
-                </div>
-              </div>
-
-              {error && (
-                <Alert variant="destructive" className="py-2">
-                  <AlertCircleIcon />
-                  <AlertDescription>{error}</AlertDescription>
-                </Alert>
-              )}
-              {success && (
-                <Alert variant="success" className="py-2">
-                  <CheckCircleIcon />
-                  <AlertDescription>{success}</AlertDescription>
-                </Alert>
-              )}
-
-              <div className="flex justify-end">
-                <Button
-                  type="submit"
-                  disabled={isLoading}
-                  className="flex items-center gap-2"
-                >
-                  <SaveIcon className="size-4" />
-                  {isLoading ? "Chargement..." : "Mettre à jour l'email"}
-                </Button>
-              </div>
-            </form>
-            </TabsContent>
-            
-            <TabsContent value="password" className="mt-0">
-            <form onSubmit={handleUpdatePassword} className="space-y-4">
-              <div className="space-y-2">
-                <Label htmlFor="currentPassword">Mot de passe actuel</Label>
-                <div className="relative">
-                  <Input
-                    id="currentPassword"
-                    type={showCurrentPassword ? "text" : "password"}
-                    value={currentPassword}
-                    onChange={(e) => setCurrentPassword(e.target.value)}
-                    placeholder="••••••••"
-                    required
-                  />
-                  <button
-                    type="button"
-                    className="absolute right-3 top-1/2 -translate-y-1/2 text-gray-500 hover:text-gray-700 focus:outline-none"
-                    onClick={() => setShowCurrentPassword(!showCurrentPassword)}
-                    tabIndex={-1}
-                  >
-                    {showCurrentPassword ? (
-                      <EyeIcon className="size-4" />
-                    ) : (
-                      <EyeOffIcon className="size-4" />
-                    )}
-                  </button>
-                </div>
-              </div>
-
-              <div className="space-y-2">
-                <Label htmlFor="newPassword">Nouveau mot de passe</Label>
-                <div className="relative">
-                  <Input
-                    id="newPassword"
-                    type={showNewPassword ? "text" : "password"}
-                    value={newPassword}
-                    onChange={(e) => setNewPassword(e.target.value)}
-                    placeholder="••••••••"
-                    required
-                  />
-                  <button
-                    type="button"
-                    className="absolute right-3 top-1/2 -translate-y-1/2 text-gray-500 hover:text-gray-700 focus:outline-none"
-                    onClick={() => setShowNewPassword(!showNewPassword)}
-                    tabIndex={-1}
-                  >
-                    {showNewPassword ? (
-                      <EyeIcon className="size-4" />
-                    ) : (
-                      <EyeOffIcon className="size-4" />
-                    )}
-                  </button>
-                </div>
-              </div>
-
-              <div className="space-y-2">
-                <Label htmlFor="confirmPassword">
-                  Confirmer le nouveau mot de passe
-                </Label>
-                <div className="relative">
-                  <Input
-                    id="confirmPassword"
-                    type={showConfirmPassword ? "text" : "password"}
-                    value={confirmPassword}
-                    onChange={(e) => setConfirmPassword(e.target.value)}
-                    placeholder="••••••••"
-                    required
-                  />
-                  <button
-                    type="button"
-                    className="absolute right-3 top-1/2 -translate-y-1/2 text-gray-500 hover:text-gray-700 focus:outline-none"
-                    onClick={() => setShowConfirmPassword(!showConfirmPassword)}
-                    tabIndex={-1}
-                  >
-                    {showConfirmPassword ? (
-                      <EyeIcon className="size-4" />
-                    ) : (
-                      <EyeOffIcon className="size-4" />
-                    )}
-                  </button>
-                </div>
-              </div>
-
-              {error && (
-                <Alert variant="destructive" className="py-2">
-                  <AlertCircleIcon />
-                  <AlertDescription>{error}</AlertDescription>
-                </Alert>
-              )}
-              {success && (
-                <Alert variant="success" className="py-2">
-                  <CheckCircleIcon />
-                  <AlertDescription>{success}</AlertDescription>
-                </Alert>
-              )}
-
-              <div className="flex justify-end">
-                <Button
-                  type="submit"
-                  disabled={isLoading}
-                  className="flex items-center gap-2"
-                >
-                  <SaveIcon className="size-4" />
-                  {isLoading
-                    ? "Chargement..."
-                    : "Mettre à jour le mot de passe"}
-                </Button>
-              </div>
-            </form>
-            </TabsContent>
-            
-            <TabsContent value="delete" className="mt-0">
-            <div className="space-y-4">
-              <Alert variant="destructive" className="py-2">
-                <AlertCircleIcon />
-                <AlertDescription>
-                  Attention : La suppression de vos données est irréversible. Toutes vos données sur Supabase seront définitivement supprimées.
-                </AlertDescription>
-              </Alert>
-              
-              <Alert variant="info" className="py-2">
-                <InfoIcon />
-                <AlertDescription>
-                  Vos données locales seront conservées. L'application passera en mode non synchronisé et vous pourrez continuer à l'utiliser sans compte.
-                </AlertDescription>
-              </Alert>
-              
-              <div className="flex justify-end">
-                <Button
-                  type="button"
-                  variant="destructive"
-                  onClick={() => setIsDeleteDialogOpen(true)}
-                  className="flex items-center gap-2"
-                >
-                  <TrashIcon className="size-4" />
-                  Supprimer mes données
-                </Button>
-              </div>
-            </div>
-            </TabsContent>
-          </Tabs>
-          
-          <AlertDialog open={isDeleteDialogOpen} onOpenChange={setIsDeleteDialogOpen}>
-            <AlertDialogContent>
-              <AlertDialogHeader>
-                <AlertDialogTitle>Êtes-vous sûr de vouloir supprimer vos données ?</AlertDialogTitle>
-                <AlertDialogDescription>
-                  Cette action est irréversible. Toutes vos données sur Supabase seront définitivement supprimées et vous serez déconnecté.
-                </AlertDialogDescription>
-              </AlertDialogHeader>
-              
-              <div className="space-y-4">
-                <Alert variant="info" className="py-2">
-                  <InfoIcon />
-                  <AlertDescription>
-                    Vos données locales seront conservées. L'application passera en mode non synchronisé.
-                  </AlertDescription>
-                </Alert>
-                
-                <div className="space-y-2">
-                  <Label htmlFor="deletePassword">Entrez votre mot de passe pour confirmer</Label>
-                  <div className="relative">
-                    <Input
-                      id="deletePassword"
-                      type={showDeletePassword ? "text" : "password"}
-                      value={deletePassword}
-                      onChange={(e) => setDeletePassword(e.target.value)}
-                      placeholder="••••••••"
-                      required
-                    />
-                    <button
-                      type="button"
-                      className="absolute right-3 top-1/2 -translate-y-1/2 text-gray-500 hover:text-gray-700 focus:outline-none"
-                      onClick={() => setShowDeletePassword(!showDeletePassword)}
-                      tabIndex={-1}
-                    >
-                      {showDeletePassword ? (
-                        <EyeIcon className="size-4" />
-                      ) : (
-                        <EyeOffIcon className="size-4" />
-                      )}
-                    </button>
-                  </div>
-                </div>
-
-                <div className="space-y-2">
-                  <Label htmlFor="deleteConfirmation">
-                    Pour confirmer, saisissez :<span className="font-semibold">"Supprimer mes données"</span>
-                  </Label>
-                  <Input
-                    id="deleteConfirmation"
-                    type="text"
-                    value={deleteConfirmation}
-                    onChange={(e) => setDeleteConfirmation(e.target.value)}
-                    placeholder="Supprimer mes données"
-                    required
-                  />
-                </div>
-                
-                {deleteError && (
-                  <Alert variant="destructive" className="py-2">
-                    <AlertCircleIcon />
-                    <AlertDescription>{deleteError}</AlertDescription>
-                  </Alert>
-                )}
-              </div>
-              
-              <AlertDialogFooter>
-                <AlertDialogCancel disabled={isLoading}>Annuler</AlertDialogCancel>
-                <AlertDialogAction
-                  onClick={(e) => {
-                    e.preventDefault();
-                    handleDeleteAccount();
-                  }}
-                  disabled={isLoading}
-                  className="bg-destructive hover:bg-destructive/90"
-                >
-                  {isLoading ? "Suppression..." : "Supprimer mes données"}
-                </AlertDialogAction>
-              </AlertDialogFooter>
-            </AlertDialogContent>
-          </AlertDialog>
-        </DialogContent>
-      </Dialog>
-    </>
-  );
-};
-
-export default ProfileButton;
